refactor: move windows from input, documentation
