+++ clippy.toml
@@ -1,0 +1,1 @@
+doc-valid-idents = ["HEx", "GHex", "CNTRLq", ".."]
